Add a load more scripting for working with MusicBrainz
The following are dead ends I'm committing for potential future reference:

- merge_files.py
- musicbrainz-api.py
- musicbrainz-ids.py
- musicbrainz-recording-json.py

The Rust app I've added replaces musicbrainz-db for matching Spotify tracks to MusicBrainz IDs. It's a lot faster than the musicbrainz-ids script, though they work in similar ways (except the script has the ability to query the DB, but that's actually slower).