Revert to unmerged recordings and serial track normalisation
It turns out that the preparation time cost of merging recordings is greater than the matching time saved, and merging recordings doesn't parallelise as well as preparing unmerged recordings. The number of tracks is also small enough that preparing them in serial is slightly faster than in parallel.

This also improves the parallel executor to preserve the order of results, as while it doesn't matter in normal usage, when I was parallel-preparing tracks and then using a subslice of them, that slice would be different depending on which thread finished its processing first.